chore: abstract carousel into seperate logo carousel file for resuse
